requirements.txt
certifi==2020.6.20
chardet==3.0.4
coverage==5.1
idna==2.9
Pillow==7.1.2
pony==0.7.13
psycopg2==2.8.5
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
time-machine==1.1.1
urllib3==1.25.9
vk-api==11.8.0
